# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Askito Rest api",
      "description": "A community for developers to share their knowledge"
    },
    "host": "askito.herokuapp.com",
    "basePath": "/api/",
...
    "schemes": ["https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
      "/questions/admin/update/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update Question",
          "description": "",
          "operationId": "updateQuestionById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Question object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Question"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/admin/questions/edit/title/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update Question title by id",
          "description": "",
          "operationId": "updateQuestionTitleById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "title",
              "description": "Question title that needs to be updated",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/admin/questions/edit/data/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update Question data by id",
          "description": "",
          "operationId": "updateQuestionDataById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "title",
              "description": "Question boby that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/admin/questions/edit/tags/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update Question tags by id",
          "description": "",
          "operationId": "updateQuestionTagsById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "tags",
              "description": "tags array",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/admin/questions/delete/{id}": {
        "delete": {
          "tags": ["Admin"],
          "summary": "delete Question by id",
          "description": "",
          "operationId": "deleteQuestionById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Question deleted"
            },
            "401": {
              "description": "access denied."
            },
            "404": {
              "description": "The Question with the given ID was not found."
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
      "/questions/add": {
        "post": {
          "tags": ["Questions"],
          "summary": "add new Question",
          "description": "",
          "operationId": "addQuestion",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Question object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Question"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Question created successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/get/{id}": {
        "get": {
          "tags": ["Questions"],
          "summary": "Get Question by id",
          "description": "",
          "operationId": "getQuestionById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be fetched",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Question"
              }
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/get/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Questions"],
          "summary": "Get total Questions",
          "description": "",
          "operationId": "getQuestions",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 posts per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Question"
              }
            },
            "404": {
              "description": "No Questions found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/get/myquestions/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Questions"],
          "summary": "Get total Question of the login user",
          "description": "",
          "operationId": "getQuestionOfLoginUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 Question per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Question"
              }
            },
            "404": {
              "description": "No Question found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/update/{id}": {
        "put": {
          "tags": ["Questions"],
          "summary": "update Question",
          "description": "",
          "operationId": "updateQuestion",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Question object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Question"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/edit/title/{id}": {
        "put": {
          "tags": ["Questions"],
          "summary": "update Question title by id",
          "description": "",
          "operationId": "updateQuestionTitleById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "title",
              "description": "Question title that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/edit/data/{id}": {
        "put": {
          "tags": ["Questions"],
          "summary": "update Question data by id",
          "description": "",
          "operationId": "updateQuestionDataById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "title",
              "description": "Question boby that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/edit/tags/{id}": {
        "put": {
          "tags": ["Questions"],
          "summary": "update Question tags by id",
          "description": "",
          "operationId": "updateQuestionTagsById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "tags",
              "description": "tags array",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Question doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/questions/delete/{id}": {
        "delete": {
          "tags": ["Questions"],
          "summary": "delete Question by id",
          "description": "",
          "operationId": "deleteQuestionById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Question to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Question deleted"
            },
            "401": {
              "description": "access denied."
            },
            "404": {
              "description": "The Question with the given ID was not found."
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
    },
      "definitions": {
...
        "Question": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "data": {
              "type": "string"
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
...
      } 
}